@@ -22,24 +22,32 @@ Map<String, dynamic> _$$_OutdatedPackagesToJson(_$_OutdatedPackages instance) =>
 _$_OutdatedPackage _$$_OutdatedPackageFromJson(Map json) => _$_OutdatedPackage(
       package: json['package'] as String,
       isDiscontinued: json['isDiscontinued'] as bool,
-      current: PackageResolution.fromJson(
-          Map<String, dynamic>.from(json['current'] as Map)),
-      upgradable: PackageResolution.fromJson(
-          Map<String, dynamic>.from(json['upgradable'] as Map)),
-      resolvable: PackageResolution.fromJson(
-          Map<String, dynamic>.from(json['resolvable'] as Map)),
-      latest: PackageResolution.fromJson(
-          Map<String, dynamic>.from(json['latest'] as Map)),
+      current: json['current'] == null
+          ? null
+          : PackageResolution.fromJson(
+              Map<String, dynamic>.from(json['current'] as Map)),
+      upgradable: json['upgradable'] == null
+          ? null
+          : PackageResolution.fromJson(
+              Map<String, dynamic>.from(json['upgradable'] as Map)),
+      resolvable: json['resolvable'] == null
+          ? null
+          : PackageResolution.fromJson(
+              Map<String, dynamic>.from(json['resolvable'] as Map)),
+      latest: json['latest'] == null
+          ? null
+          : PackageResolution.fromJson(
+              Map<String, dynamic>.from(json['latest'] as Map)),
     );
 
 Map<String, dynamic> _$$_OutdatedPackageToJson(_$_OutdatedPackage instance) =>
     <String, dynamic>{
       'package': instance.package,
       'isDiscontinued': instance.isDiscontinued,
-      'current': instance.current.toJson(),
-      'upgradable': instance.upgradable.toJson(),
-      'resolvable': instance.resolvable.toJson(),
-      'latest': instance.latest.toJson(),
+      'current': instance.current?.toJson(),
+      'upgradable': instance.upgradable?.toJson(),
+      'resolvable': instance.resolvable?.toJson(),
+      'latest': instance.latest?.toJson(),
     };
 
 _$_PackageResolution _$$_PackageResolutionFromJson(Map json) =>
